+++ model/transfer/transfer-model.js
@@ -10,13 +10,43 @@ const Model = require('../../lib/model');
 /**
  * @typedef {Object} TransferData
  * @property {TransferId} _id -
- * @property {ManagerId} ownerId -
- * @property {PlayerId} inboundPlayer -
- * @property {PlayerId} outboundPlayer -
+ * @property {ManagerId} managerId -
+ * @property {PlayerId} inbound -
+ * @property {PlayerId} outbound -
  * @property {number} priority -
  */
 
 class TransferModel extends Model {
+
+    /**
+     * Loads a transfer from a database.
+     * @param {LoggingContext} ctx -
+     * @param {{ managerId: ManagerId, priority: number }} query -
+     * @returns {Promise<TransferData>} - 
+     */
+    async load(ctx, { managerId, priority }) {
+        try {
+            this.logDebug(ctx, 'Loading a transfer from a database.');
+
+            const transfer = await this._storage.get(ctx, {
+                query: {
+                    managerId,
+                    priority,
+                },
+                collection: this._collection,
+            });
+
+            this.logDebug(ctx, 'A transfer is obtained from a database.', {
+                transfer,
+            });
+
+            return transfer;
+        } catch (error) {
+            this.logError(ctx, 'Error on loading a transfer from a database.', { error });
+
+            throw error;
+        }
+    }
 
     /**
      * Loads multiple transfers from a database.
@@ -104,6 +134,54 @@ class TransferModel extends Model {
             return transfer;
         } catch (error) {
             this.logError(ctx, 'Error on creating a transfer.', { error });
+
+            return error;
+        }
+    }
+
+    /**
+     * Deletss a transfer of a manager and repriorized rest of them.
+     * @param {LoggingContext} ctx -
+     * @param {Object} options - 
+     * @param {ManagerId} options.managerId -
+     * @param {number} options.priority -
+     * @returns {Promise<void>} -
+     */
+    async deleteAndReprioritize(ctx, { managerId, priority }) {
+        try {
+            this.logDebug(ctx, 'Deleting a transfer from a database.', {
+                managerId, priority,
+            });
+
+            // TODO: Rework thay layer so MongoDB won't appear here.
+            await this._storage.bulkUpdate(ctx, [{
+                deleteOne: {
+                    filter: {
+                        managerId,
+                        priority,
+                    },
+                },
+            }, {
+                updateMany: {
+                    filter: {
+                        managerId,
+                        priority: {
+                            $gt: priority,
+                        }
+                    },
+                    update: {
+                        $inc: {
+                            priority: -1,
+                        },
+                    },
+                },
+            }], {
+                collection: this._collection,
+            });
+
+            this.logDebug(ctx, 'A transfer is deleted from a database.');
+        } catch (error) {
+            this.logError(ctx, 'Error on deleting a transfer.', { error });
 
             return error;
         }